Changed to teh generated light definitions (see: https://github.com/SiebenCorgie/ori-engine/issues/17 )
At least for lights, will have a look if this might be extensible for all the other structs.
Might be worth it though. (save and stuff).

Will also have to investigate how to create a somwhat dynamic light cap. Currently the arrays are
sized to 6 lights per type max. However, thats not dynamic. Will investigate.


Former-commit-id: badf3107cf35e2e26fa580f328ba80f1eddc9ae8